+++ nata-travel/src/index.js
@@ -1,20 +1,10 @@
-import adb from 'adbkit'
+import Monkey from './Monkey'
 
-const bluebird = require('bluebird')
-const client = adb.createClient()
+class DFSMonkey extends Monkey {
 
-client.listDevices()
-  .then((devices) => {
-    return bluebird.filter(devices, (device) => {
-      return client.getFeatures(device.id)
-        .then((features) => {
-          return features['android.hardware.nfc']
-        })
-    })
-  })
-  .then((supportedDevices) => {
-    console.log('The following devices support NFC:', supportedDevices)
-  })
-  .catch((err) => {
-    console.error('Something went wrong:', err.stack)
-  })
+  play() {
+
+  }
+}
+
+export default DFSMonkey